Use Java 16 & latest Spring & remove lombok
Lombok was primarily being used because it integrated so easily into the
build process since it is an annotation processor but since it uses
internal APIs using it in the future will be more difficult. However,
because Java 16's records features fulfill most of what I used lombok
for (immutable value classes) I upgraded to Java 16 and removed lombok.