Remove database module
This project was intended to be a super straight forward "template" to
use for other projects. Jooq and Flyway are not super straight forward
to use with Gradle and would likely have special requirements instead of
being used the same way in each project (like Spring Boot). (In addition
they're more opinionated choices than Spring Boot.)